fix(gzip): check if files exists before create read/write stream

+++ daily-rotate-file.js
@@ -97,17 +97,15 @@ var DailyRotateFile = function (options) {
         if (options.zippedArchive) {
             this.logStream.on('rotate', function (oldFile) {
                 var gzip = zlib.createGzip();
-                var inp = fs.createReadStream(oldFile);
-                inp.on('error', function () {
-                    var zipExist = fs.createReadStream(oldFile + '.gz');
-                    zipExist.on('error', function () {
-                        throw new Error('File ' + oldFile + 'does not exist');
+                var oldFileExist = fs.existsSync(oldFile);
+                var gzExist = fs.existsSync(oldFile + '.gz');
+                if (oldFileExist && !gzExist) {
+                    var inp = fs.createReadStream(oldFile);
+                    var out = fs.createWriteStream(oldFile + '.gz');
+                    inp.pipe(gzip).pipe(out).on('finish', function () {
+                        fs.unlinkSync(oldFile);
                     });
-                });
-                var out = fs.createWriteStream(oldFile + '.gz');
-                inp.pipe(gzip).pipe(out).on('finish', function () {
-                    fs.unlinkSync(oldFile);
-                });
+                }
             });
         }
     }
